[hotfix][test] Separate some shared inner classes from test classes into the separate files
